把結帳邏輯從cart切出來
